sorted by maxOccupancy
