+++ .idea/sonarIssues.xml
@@ -108,7 +108,17 @@
             <set />
           </value>
         </entry>
+        <entry key="$PROJECT_DIR$/app/src/main/java/com/luckytrip/luckytrip/extenstions/Room.kt">
+          <value>
+            <set />
+          </value>
+        </entry>
         <entry key="$PROJECT_DIR$/app/src/main/java/com/luckytrip/luckytrip/extenstions/String.kt">
+          <value>
+            <set />
+          </value>
+        </entry>
+        <entry key="$PROJECT_DIR$/app/src/main/java/com/luckytrip/luckytrip/json/KoinInit.kt">
           <value>
             <set />
           </value>
